refactor pd core tests
