+++ hugegraph-pd/hg-pd-test/src/main/java/org/apache/hugegraph/pd/core/TaskScheduleServiceTest.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.hugegraph.pd.service;
+package org.apache.hugegraph.pd.core;
 
 import static org.junit.Assert.assertTrue;
 
@@ -28,7 +28,7 @@ import org.apache.hugegraph.pd.grpc.Metapb;
 import org.junit.Before;
 import org.junit.Test;
 
-public class TaskScheduleServiceTest extends PdTestBase {
+public class TaskScheduleServiceTest extends PDCoreTestBase {
 
     TaskScheduleService service;
