+++ hugegraph-pd/hg-pd-test/src/main/java/org/apache/hugegraph/pd/rest/PDRestSuiteTest.java
@@ -1,0 +1,16 @@
+package org.apache.hugegraph.pd.rest;
+
+import org.junit.runner.RunWith;
+import org.junit.runners.Suite;
+
+import lombok.extern.slf4j.Slf4j;
+
+@RunWith(Suite.class)
+@Suite.SuiteClasses({
+        RestApiTest.class,
+})
+
+@Slf4j
+public class PDRestSuiteTest {
+
+}
